change pro platform to rsc and streamline auth
